feat: Refactor event modules and add Transfer struct
Refactored the event modules `l1.rs` and `l2.rs` to use a common `Transfer` struct for deposit and withdrawal events. This improves code organization and reduces duplication. Also added the `Transfer` struct to the `mod.rs` file for shared usage.
