+++ src/events/l2.rs
@@ -1,4 +1,8 @@
 use serde::{Deserialize, Serialize};
 
+use super::Transfer;
+
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub enum L2Event {}
+pub enum L2Event {
+    Withdrawal(Transfer),
+}
